src/brdf/brdf.pro: Tell users how to define $prefix
`pf-makevar` does not always exist so inform the user about how to
define the installation prefix.  Pipe errors to /dev/null to avoid
console noise.

Default LIBDIR to "lib" since that is the most common scenario.
Check to see if LIBDIR is defined before setting it so that users can
provide their own value if needed.

Signed-off-by: David Aguilar <david.aguilar@disneyanimation.com>